add prisma/dev.sqlite* to .dockerignore
The current Dockerfile will fail unless you've kept the default
provider (sqlite) and run npx prisma migrate dev locally.  If you
switch to another database or try to run the Dockerfile without having
first set up a local database, your build will fail.

A better solution is to not include the sqlite datbase in the first place.
This will save upload time and will work whether or not the database is
present.